repo_results_2/run-20240312-153149_LLMatic-ZeroShot-CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-twitter-donnemartin-system-design-oop-whatsapp_0_20-runs_log/twitter/4/LLMatic/trending.py
class Trending:
	def __init__(self):
		self.trending_topics = {}

	def get_trending_topics(self):
		# Sort the trending topics based on volume and velocity of mentions
		trending = sorted(self.trending_topics.items(), key=lambda x: x[1], reverse=True)
		return trending

	def sort_trending_topics(self, location=None):
		# If location is specified, sort the trending topics based on location
		if location:
			trending = {k: v for k, v in self.trending_topics.items() if k.startswith(location)}
			return sorted(trending.items(), key=lambda x: x[1], reverse=True)
		# If location is not specified, return the global trending topics
		return self.get_trending_topics()

